svn: re-factor svn design
1. re-factor get_ufm_svn and set_ufm_svn API and change its data type of offset parameter to unit32_t
2. change SVN_MAX to 64 instead of 63
3. remove reduant PFM structure. replace _PFM_STRUCTURE_1 with _PFM_STRUCTURE structure

+++ apps/aspeed-pfr/src/cerberus_pfr/cerberus_pfr_update.c
@@ -35,7 +35,7 @@ int cerberus_set_ufm_svn(struct pfr_manifest *manifest, uint8_t ufm_location, ui
 	return Success;
 }
 
-int get_ufm_svn(struct pfr_manifest *manifest, uint8_t offset)
+uint8_t get_ufm_svn(struct pfr_manifest *manifest, uint32_t offset)
 {
 	return 0;
 }
